move pages to IDP specific packages, and add help info to the google profile
