+++ src/test/resources/application-google.yml
@@ -1,30 +1,19 @@
-info:
-  name: ${project.artifactId}
-  version: ${project.version}
-  origin: ${scm.url}
-  commit: ${buildNumber}
-  branch: ${scmBranch}
-  built: ${timestamp}
-
-# logging config here as a reminder as to how to override logback from cmd line
-# ie -Dlogging.level.org.springframework=DEBUG
-logging:
-  level:
-    root: INFO
-    uk.gov.ons.ctp: INFO
-    org.springframework: INFO
-  profile: DEV
-  useJson: false
+#
+# You can use this google profile, if the field service is running against google IDP
+# and it you are running locally.
+#
+# The reason why you must be running locally is because on the kubernetes environment 
+# you will get a Captcha which will fail the cucumber tests.
+#
+# An example of running with this profile locally:
+#
+#   export CONFIG_PASSWORD=<the password for the user-id below>
+#
+#   mvn clean install -Dspring.profiles.active=google
+#
 
 config:
   idp-type: google
   user-id: fieldservice_test@test.field.census.gov.uk
   password: override_me_with_runtime_property
-  base-url: https://localhost:443
-  eq-host: www.google.com
-
-webdriver:
-  type: FIREFOX
-  logging_level: WARN
-  headless: true
